+++ app/src/main/java/com/example/payme20/model/IDebtUpdater.java
@@ -4,5 +4,5 @@ import java.util.List;
 import java.util.Map;
 
 public interface IDebtUpdater {
-    List<Debt> updateDebts(Map<Member, Double> map, Member Payer);
+    List<Debt> updateDebts(Map<Member, Integer> map, Member Payer);
 }
