+++ app/src/main/java/com/example/payme20/model/Model.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public class Model {
 
 
-    public void createNewGroupEvent(Group group, Map<Member, Double> debtMap, String eventName, Member payer, IDebtUpdater iDebtUpdater) {
+    public void createNewGroupEvent(Group group, Map<Member, Integer> debtMap, String eventName, Member payer, IDebtUpdater iDebtUpdater) {
         Event event = Factory.createEvent(eventName, debtMap, payer, iDebtUpdater);
         group.addEvent(event);
     }
