+++ app/src/main/java/com/example/payme20/model/SplitDebtUpdater.java
@@ -6,30 +6,30 @@ import java.util.Map;
 
 public class SplitDebtUpdater implements IDebtUpdater {
     @Override
-    public List<Debt> updateDebts(Map<Member, Double> eventMemberPaidAmount, Member payer) {
-        double totalGroupCost = calcTotalGroupCost(eventMemberPaidAmount);
-        double splitCost = calcDividedCost(totalGroupCost, eventMemberPaidAmount.size());
+    public List<Debt> updateDebts(Map<Member, Integer> eventMemberPaidAmount, Member payer) {
+        int totalGroupCost = calcTotalGroupCost(eventMemberPaidAmount);
+        int splitCost = calcDividedCost(totalGroupCost, eventMemberPaidAmount.size());
         return createEventDebtList(eventMemberPaidAmount, payer, splitCost);
     }
 
-    private double calcTotalGroupCost(Map<Member, Double> eventMemberPaidAmount) {
-        double total = 0;
-        for (Map.Entry<Member, Double> mapValue: eventMemberPaidAmount.entrySet()) {
+    private int calcTotalGroupCost(Map<Member, Integer> eventMemberPaidAmount) {
+        int total = 0;
+        for (Map.Entry<Member, Integer> mapValue: eventMemberPaidAmount.entrySet()) {
             total += mapValue.getValue();
         }
         return total;
     }
 
-    private double calcDividedCost(double totalGroupCost, int memberSize){
-        double dividedCost = 0;
+    private int calcDividedCost(int totalGroupCost, int memberSize){
+        int dividedCost = 0;
         try{dividedCost = totalGroupCost / memberSize;}
         catch (ArithmeticException e){System.out.println("Division by zero in SplitDebtUpdater.Class");}
         return dividedCost;
     }
 
-    private List<Debt> createEventDebtList(Map<Member, Double> eventMemberPaidAmount, Member memberToGetPaid, double debtAmount){
+    private List<Debt> createEventDebtList(Map<Member, Integer> eventMemberPaidAmount, Member memberToGetPaid, int debtAmount){
         List<Debt> eventDebtList = new ArrayList<>();
-        for (Map.Entry<Member, Double> debtMap: eventMemberPaidAmount.entrySet()) {
+        for (Map.Entry<Member, Integer> debtMap: eventMemberPaidAmount.entrySet()) {
             Member memberToPay = debtMap.getKey();
             if(!(memberToPay.equals(memberToGetPaid))){
                 eventDebtList.add(new Debt(memberToGetPaid, memberToPay, debtAmount));
